Properties Query and Collection Members Query refactored
